# update trie dependency to take advantage of improvements