Closes #38