@prologic 2019-08-05
On my 11" Macbook:

Before:

```
goos: darwin
goarch: amd64
pkg: github.com/prologic/bitcask
BenchmarkGet/128B-4         	  300000	      4387 ns/op	  29.17 MB/s	     387 B/op	       4 allocs/op
BenchmarkGet/128BWithPool-4 	  300000	      5170 ns/op	  24.76 MB/s	     227 B/op	       3 allocs/op
BenchmarkGet/256B-4         	  300000	      4759 ns/op	  53.79 MB/s	     643 B/op	       4 allocs/op
BenchmarkGet/256BWithPool-4 	  300000	      4702 ns/op	  54.44 MB/s	     355 B/op	       3 allocs/op
BenchmarkGet/512B-4         	  300000	      5087 ns/op	 100.64 MB/s	    1187 B/op	       4 allocs/op
BenchmarkGet/512BWithPool-4 	  200000	      5353 ns/op	  95.65 MB/s	     611 B/op	       3 allocs/op
BenchmarkGet/1K-4           	  200000	      5726 ns/op	 178.83 MB/s	    2275 B/op	       4 allocs/op
BenchmarkGet/1KWithPool-4   	  300000	      5749 ns/op	 178.10 MB/s	    1123 B/op	       3 allocs/op
BenchmarkGet/2K-4           	  200000	      6100 ns/op	 335.69 MB/s	    4451 B/op	       4 allocs/op
BenchmarkGet/2KWithPool-4   	  200000	      6279 ns/op	 326.16 MB/s	    2147 B/op	       3 allocs/op
BenchmarkGet/4K-4           	  200000	      7463 ns/op	 548.80 MB/s	    9059 B/op	       4 allocs/op
BenchmarkGet/4KWithPool-4   	  200000	      7119 ns/op	 575.34 MB/s	    4195 B/op	       3 allocs/op
BenchmarkGet/8K-4           	  200000	     10232 ns/op	 800.61 MB/s	   17763 B/op	       4 allocs/op
BenchmarkGet/8KWithPool-4   	  200000	      9420 ns/op	 869.59 MB/s	    8291 B/op	       3 allocs/op
BenchmarkGet/16K-4          	  100000	     14019 ns/op	1168.67 MB/s	   34915 B/op	       4 allocs/op
BenchmarkGet/16KWithPool-4  	  100000	     12118 ns/op	1351.96 MB/s	   16483 B/op	       3 allocs/op
BenchmarkGet/32K-4          	  100000	     22883 ns/op	1431.95 MB/s	   73827 B/op	       4 allocs/op
BenchmarkGet/32KWithPool-4  	  100000	     18062 ns/op	1814.15 MB/s	   32867 B/op	       3 allocs/op
BenchmarkPut/128B-4         	  100000	     12951 ns/op	   9.88 MB/s	     409 B/op	       6 allocs/op
BenchmarkPut/256B-4         	  100000	     14670 ns/op	  17.45 MB/s	     539 B/op	       6 allocs/op
BenchmarkPut/512B-4         	  100000	     21893 ns/op	  23.39 MB/s	     829 B/op	       6 allocs/op
BenchmarkPut/1K-4           	   50000	     28981 ns/op	  35.33 MB/s	    1411 B/op	       6 allocs/op
BenchmarkPut/2K-4           	   30000	     38122 ns/op	  53.72 MB/s	    2574 B/op	       6 allocs/op
BenchmarkPut/4K-4           	   20000	     72182 ns/op	  56.75 MB/s	    5155 B/op	       6 allocs/op
BenchmarkPut/8K-4           	   10000	    126955 ns/op	  64.53 MB/s	    9813 B/op	       6 allocs/op
BenchmarkPut/16K-4          	   10000	    195128 ns/op	  83.97 MB/s	   18852 B/op	       6 allocs/op
BenchmarkPut/32K-4          	    5000	    354576 ns/op	  92.41 MB/s	   41549 B/op	       7 allocs/op
BenchmarkScan-4             	 1000000	      2202 ns/op	     493 B/op	      25 allocs/op
PASS
ok  	github.com/prologic/bitcask	51.915s

```

After:

```
goos: darwin
goarch: amd64
pkg: github.com/prologic/bitcask
BenchmarkGet/128B-4         	  300000	      3915 ns/op	  32.69 MB/s	     387 B/op	       4 allocs/op
BenchmarkGet/128BWithPool-4 	  300000	      4407 ns/op	  29.04 MB/s	     227 B/op	       3 allocs/op
BenchmarkGet/256B-4         	  300000	      4251 ns/op	  60.21 MB/s	     643 B/op	       4 allocs/op
BenchmarkGet/256BWithPool-4 	  300000	      4915 ns/op	  52.08 MB/s	     355 B/op	       3 allocs/op
BenchmarkGet/512B-4         	  300000	      4832 ns/op	 105.94 MB/s	    1187 B/op	       4 allocs/op
BenchmarkGet/512BWithPool-4 	  300000	      5354 ns/op	  95.62 MB/s	     611 B/op	       3 allocs/op
BenchmarkGet/1K-4           	  200000	      5560 ns/op	 184.17 MB/s	    2275 B/op	       4 allocs/op
BenchmarkGet/1KWithPool-4   	  200000	      5525 ns/op	 185.32 MB/s	    1123 B/op	       3 allocs/op
BenchmarkGet/2K-4           	  200000	      5974 ns/op	 342.77 MB/s	    4451 B/op	       4 allocs/op
BenchmarkGet/2KWithPool-4   	  200000	      6095 ns/op	 335.97 MB/s	    2147 B/op	       3 allocs/op
BenchmarkGet/4K-4           	  200000	      7408 ns/op	 552.85 MB/s	    9059 B/op	       4 allocs/op
BenchmarkGet/4KWithPool-4   	  200000	      7311 ns/op	 560.19 MB/s	    4195 B/op	       3 allocs/op
BenchmarkGet/8K-4           	  200000	     10637 ns/op	 770.13 MB/s	   17763 B/op	       4 allocs/op
BenchmarkGet/8KWithPool-4   	  200000	      9344 ns/op	 876.69 MB/s	    8291 B/op	       3 allocs/op
BenchmarkGet/16K-4          	  100000	     14054 ns/op	1165.74 MB/s	   34915 B/op	       4 allocs/op
BenchmarkGet/16KWithPool-4  	  100000	     12056 ns/op	1358.91 MB/s	   16483 B/op	       3 allocs/op
BenchmarkGet/32K-4          	   50000	     23071 ns/op	1420.27 MB/s	   73827 B/op	       4 allocs/op
BenchmarkGet/32KWithPool-4  	  100000	     18331 ns/op	1787.57 MB/s	   32867 B/op	       3 allocs/op
BenchmarkPut/128B-4         	  100000	     13015 ns/op	   9.83 MB/s	     441 B/op	       6 allocs/op
BenchmarkPut/256B-4         	  100000	     14433 ns/op	  17.74 MB/s	     571 B/op	       6 allocs/op
BenchmarkPut/512B-4         	  100000	     19758 ns/op	  25.91 MB/s	     861 B/op	       6 allocs/op
BenchmarkPut/1K-4           	   50000	     29125 ns/op	  35.16 MB/s	    1442 B/op	       6 allocs/op
BenchmarkPut/2K-4           	   30000	     39470 ns/op	  51.89 MB/s	    2606 B/op	       6 allocs/op
BenchmarkPut/4K-4           	   20000	     75099 ns/op	  54.54 MB/s	    5187 B/op	       6 allocs/op
BenchmarkPut/8K-4           	   10000	    127651 ns/op	  64.17 MB/s	    9845 B/op	       6 allocs/op
BenchmarkPut/16K-4          	   10000	    200842 ns/op	  81.58 MB/s	   18883 B/op	       6 allocs/op
BenchmarkPut/32K-4          	    5000	    347692 ns/op	  94.24 MB/s	   41581 B/op	       7 allocs/op
BenchmarkScan-4             	 1000000	      1465 ns/op	     408 B/op	      16 allocs/op
PASS
ok  	github.com/prologic/bitcask	48.510s
```